# Optimize websocket headers handling