<!--
PLEASE READ THIS MESSAGE.

Documentation fixes or enhancements:
- for Traefik v1: use branch v1.7
- for Traefik v2: use branch v2.8

Bug fixes:
- for Traefik v1: use branch v1.7
- for Traefik v2: use branch v2.8

Enhancements:
- for Traefik v1: we only accept bug fixes
- for Traefik v2: use branch master

HOW TO WRITE A GOOD PULL REQUEST? https://doc.traefik.io/traefik/contributing/submitting-pull-requests/

-->

### What does this PR do?

Improves performance for non websocket upgrade requests.

### Motivation

Improve performance !!!!!!

### More

~- [ ] Added/updated tests~
~- [ ] Added/updated documentation~

### Additional Notes
Before
```
Running 30s test @ http://traefik/bench
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.63ms    1.11ms  44.00ms   81.79%
    Req/Sec    16.22k     1.23k   54.86k    99.17%
  Latency Distribution
     50%    1.42ms
     75%    2.07ms
     90%    2.87ms
     99%    5.03ms
  1938441 requests in 30.10s, 188.56MB read
Requests/sec:  64401.47
Transfer/sec:      6.26MB
```
After
```
Running 30s test @ http://traefik/bench
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.56ms    1.04ms  53.60ms   81.83%
    Req/Sec    16.93k   532.62    18.78k    93.25%
  Latency Distribution
     50%    1.36ms
     75%    1.97ms
     90%    2.71ms
     99%    4.79ms
  2021318 requests in 30.02s, 196.62MB read
Requests/sec:  67331.54
Transfer/sec:      6.55MB

```

Co-authored-by: Kevin Pollet <pollet.kevin@gmail.com>